requirements.txt
tensorflow=2.9.1
tensorflow-hub=0.12.0
flask=2.1.2
Pillow=9.1.1
scipy=1.8.1
python-dotenv=0.20.0
numpy=1.23.0
